more restructuring of the event modal
