+++ src/components/EventModal/EventModal.tsx
@@ -1,119 +1,21 @@
-import {
-    Button,
-    Flex,
-    Modal,
-    Stack,
-    Text,
-    TextInput,
-    useCombobox,
-} from '@mantine/core'
+import { Button, Flex, Modal, Stack, Text, TextInput } from '@mantine/core'
 import '@mantine/dates/styles.css'
 import { useDisclosure, useViewportSize } from '@mantine/hooks'
-import { DateTimePicker } from '@mantine/dates'
-import { FormEvent, useEffect, useState } from 'react'
-import supabase, { eventsAtom } from '../../lib/state'
+import { eventsAtom } from '../../lib/state'
 import { useAtom } from 'jotai'
-import { dateToPosition, dateToWidth } from '../../lib/utils'
-import { z } from 'zod'
-import {
-    FormDataschema,
-    FormDataValidateschema,
-    NewDeadlineschema,
-    NewDeadlineValidateschema,
-} from './schemas'
-import { emptyForm, formDataAtom } from './state'
-import Deadline from './SubComponents/deadline'
+import { errorsAtom, formDataAtom } from './state'
+import Deadline from './SubComponents/Deadline'
 import DropdownSelect from './SubComponents/DropdownSelect'
+import NewDeadline from './SubComponents/NewDeadline'
+import insertSupabase from './utils'
 
 const NewEventModal = () => {
-    type FromData = z.infer<typeof FormDataschema>
     const [opend, { open, close }] = useDisclosure(false)
     const [formData, setFormData] = useAtom(formDataAtom)
     const [events, setEvents] = useAtom(eventsAtom)
     const viewport = useViewportSize()
-    const [errors, setErrors] = useState({
-        title: '',
-        type: '',
-        deadlines: '',
-    })
-    const [newErrors, setNewErrors] = useState({
-        name: '',
-        timestamp: '',
-    })
-    const [newDeadline, setNewDeadline] = useState<
-        z.infer<typeof NewDeadlineschema>
-    >({
-        name: '',
-        timestamp: null,
-    })
+    const [errors, setErrors] = useAtom(errorsAtom)
 
-    const addNewDeadline = () => {
-        const parsedData = NewDeadlineValidateschema.safeParse(newDeadline)
-        let tempErrors = { name: '', timestamp: '' }
-        if (parsedData.success) {
-            setNewErrors(tempErrors)
-            setFormData({
-                ...formData,
-                deadlines: [...formData.deadlines, newDeadline],
-            })
-            setNewDeadline({
-                name: '',
-                timestamp: null,
-            })
-        } else {
-            const keys = ['name', 'timestamp']
-            parsedData.error.issues.map((issue) => {
-                keys.map((key) => {
-                    if (issue.path[0] === key) {
-                        tempErrors = { ...tempErrors, [key]: issue.message }
-                    }
-                })
-            })
-            setNewErrors(tempErrors)
-        }
-    }
-
-    const insertSupabase = async (e: FormEvent<HTMLFormElement>) => {
-        e.preventDefault()
-        let tempErrors = { title: '', type: '', deadlines: '' }
-        const parsedData = FormDataValidateschema.safeParse(formData)
-        if (parsedData.success) {
-            setErrors(tempErrors)
-            //@ts-ignore
-            const { data, error } = await supabase
-                .from('newEvent')
-                .insert({
-                    title: formData.title,
-                    type: formData.type,
-                    deadlines: JSON.stringify(formData.deadlines),
-                })
-                .select()
-                .single()
-            if (error) {
-                console.log(error)
-                return
-            }
-            setFormData(emptyForm as FromData)
-        } else {
-            const keys = ['title', 'type', 'deadlines']
-            parsedData.error.issues.map((issue) => {
-                keys.map((key) => {
-                    if (issue.path[0] === key) {
-                        tempErrors = { ...tempErrors, [key]: issue.message }
-                    }
-                })
-            })
-            setErrors(tempErrors)
-        }
-    }
-
-    useEffect(() => {
-        console.log(errors, newErrors)
-    }, [errors, newErrors])
-
-    // useEffect(() => {
-    //     console.log(formData.deadlines)
-    // }, [formData.deadlines])
     return (
         <>
             <Modal
@@ -153,45 +55,7 @@ const NewEventModal = () => {
                             {errors.deadlines !== '' && (
                                 <Text>{errors.deadlines}</Text>
                             )}
-                            <Flex align={'center'} w={'100%'}>
-                                <Flex direction={'column'} gap={6} w={'100%'}>
-                                    <TextInput
-                                        size="sm"
-                                        w={'45%'}
-                                        placeholder="Deadline Name"
-                                        value={newDeadline.name}
-                                        onChange={(e) => {
-                                            setNewDeadline({
-                                                ...newDeadline,
-                                                name: e.target.value,
-                                            })
-                                        }}
-                                        error={
-                                            newErrors.name !== ''
-                                                ? newErrors.name
-                                                : ''
-                                        }
-                                    />
-                                    <DateTimePicker
-                                        w={'35%'}
-                                        size="xs"
-                                        placeholder="2024/01/01 00:00"
-                                        value={newDeadline.timestamp}
-                                        onChange={(e) => {
-                                            setNewDeadline({
-                                                ...newDeadline,
-                                                timestamp: e as Date,
-                                            })
-                                        }}
-                                    />
-                                    {newErrors.timestamp !== '' && (
-                                        <Text>{newErrors.timestamp}</Text>
-                                    )}
-                                </Flex>
-                                <Button type="button" onClick={addNewDeadline}>
-                                    +
-                                </Button>
-                            </Flex>
+                            <NewDeadline />
                         </Stack>
                         <Button
                             type="submit"
