+++ src/components/EventModal/state.ts
@@ -5,3 +5,4 @@ import { atom } from 'jotai'
 type FromData = z.infer<typeof FormDataschema>
 export const emptyForm = { title: '', type: null, deadlines: [] }
 export const formDataAtom = atom<FromData>(emptyForm as FromData)
+export const errorsAtom = atom({ title: '', type: '', deadlines: '' })
